requirements.txt
Flask==2.1.0
Flask_Cors==3.0.10
Flask_Mail==0.9.1
imgspy==0.2.2
Pillow==9.1.0
PyJWT==2.3.0
pytest==6.2.2
requests==2.22.0
Werkzeug==2.0
